Split slackbot into two cloud functions
Splits the slackbot into two cloud functions, one works as the http
endpoint that slack will actually talk to (and responds quickly so that
the timeout is not hit) and the other works as the background worker
that actually handles the user or scheduler input.

This also merges to two scheduled tasks (purging expired whitelisted ips
and escalating alerts after their ttl) into the background worker.

Lastely, made some small fixes in other parts of the repo as I ran the
repo (except cloudtrail-streamer) through the newest go meta linter.

Fixes https://bugzilla.mozilla.org/show_bug.cgi?id=1547373
