+++ duopull/go.mod
@@ -2,10 +2,9 @@ module github.com/mozilla-services/foxsec-pipeline-contrib/duopull
 
 require (
 	cloud.google.com/go v0.37.4
-	github.com/gorilla/websocket v1.4.0 // indirect
 	github.com/mozilla-services/foxsec-pipeline-contrib v0.0.0
 	github.com/mozilla-services/foxsec-pipeline-contrib/foxsec-slack-bot v0.0.0-20190422180541-854a65bd9948 // indirect
-	github.com/sirupsen/logrus v1.3.0
+	github.com/sirupsen/logrus v1.4.1
 	go.mozilla.org/mozlogrus v1.0.1-0.20171031175137-a4ca0c1ee1cb
 )
