--- scheduled-tasks/whitelisted-ips-purger/.gcloudignore
@@ -1,2 +1,0 @@
-go.mod
-go.sum
