--- scheduled-tasks/whitelisted-ips-purger/go.mod
@@ -1,9 +1,0 @@
-module github.com/mozilla-services/foxsec-pipeline-contrib/scheduled-tasks/whitelisted-ips-purger
-
-require (
-	github.com/mozilla-services/foxsec-pipeline-contrib v0.0.0
-	github.com/sirupsen/logrus v1.3.0
-	go.mozilla.org/mozlogrus v1.0.0
-)
-
-replace github.com/mozilla-services/foxsec-pipeline-contrib v0.0.0 => ../../
