+++ slackbot-background/.gcloudignore
