+++ slackbot-background/background.go
@@ -1,9 +1,7 @@
-package foxsecslackbot
+package slackbotbackground
 
 import (
-	"bytes"
 	"context"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"strings"
@@ -12,22 +10,13 @@ import (
 	"github.com/mozilla-services/foxsec-pipeline-contrib/common"
 	"github.com/mozilla-services/foxsec-pipeline-contrib/common/persons_api"
 
+	"cloud.google.com/go/pubsub"
 	"github.com/nlopes/slack"
 	log "github.com/sirupsen/logrus"
 	"go.mozilla.org/mozlogrus"
 )
 
-var (
-	globalConfig            Config
-	client                  *http.Client
-	KEYNAME                 string
-	PROJECT_ID              string
-	FOXSEC_SLACK_CHANNEL_ID string
-	DB                      *common.DBClient
-)
-
 const (
-	EMAIL_CHAR_SET                     = "UTF-8"
 	WHITELIST_IP_SLASH_COMMAND         = "/whitelist_ip"
 	STAGING_WHITELIST_IP_SLASH_COMMAND = "/staging_whitelist_ip"
 	DEFAULT_EXPIRATION_DURATION        = time.Hour * 24
@@ -37,8 +26,18 @@ const (
 		"the default (24 hours) is used. If your duration is under 5 minutes, it is increased to 5 minutes."
 )
 
+var (
+	globalConfig            Config
+	client                  *http.Client
+	KEYNAME                 string
+	PROJECT_ID              string
+	FOXSEC_SLACK_CHANNEL_ID string
+	DB                      *common.DBClient
+	ALERT_ESCALATION_TTL    time.Duration
+)
+
 func init() {
-	mozlogrus.Enable("foxsec-slack-bot")
+	mozlogrus.Enable("slackbot-background")
 	client = &http.Client{
 		Timeout: 10 * time.Second,
 	}
@@ -56,10 +55,14 @@ func init() {
 		log.Errorf("Error creating db client: %s", err)
 		return
 	}
+
+	ALERT_ESCALATION_TTL, err = time.ParseDuration(os.Getenv("ALERT_ESCALATION_TTL"))
+	if err != nil {
+		log.Fatalf("Failed to parse alert escalation ttl: %s | Err: %s", os.Getenv("ALERT_ESCALATION_TTL"), err)
+	}
 }
 
 type Config struct {
-	slackSigningSecret  string
 	slackAuthToken      string
 	slackClient         *slack.Client
 	personsClientId     string
@@ -88,11 +91,6 @@ func InitConfig() {
 	globalConfig.sesClient, err = common.NewSESClient(os.Getenv("AWS_REGION"), accessKeyId, secretAccessKey, os.Getenv("SES_SENDER_EMAIL"), os.Getenv("ESCALATION_EMAIL"))
 	if err != nil {
 		log.Fatalf("Could not setup SESClient. Err: %s", err)
-	}
-
-	globalConfig.slackSigningSecret, err = kms.DecryptEnvVar(KEYNAME, "SLACK_SIGNING_SECRET")
-	if err != nil {
-		log.Fatalf("Could not decrypt slack signing secret. Err: %s", err)
 	}
 
 	globalConfig.slackAuthToken, err = kms.DecryptEnvVar(KEYNAME, "SLACK_AUTH_TOKEN")
@@ -125,53 +123,87 @@ func InitConfig() {
 	globalConfig.allowedGroups = strings.Split(os.Getenv("ALLOWED_LDAP_GROUPS"), ",")
 }
 
-func FoxsecSlackBot(w http.ResponseWriter, r *http.Request) {
-	w.WriteHeader(http.StatusOK)
-
-	err := verifySignature(r)
+func alertEscalator(ctx context.Context) error {
+	alerts, err := DB.GetAllAlerts(ctx)
 	if err != nil {
-		log.Errorf("Error verifying signature: %s", err)
-		return
+		log.Errorf("Error getting all alerts: %s", err)
+		return err
 	}
 
-	body, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		log.Errorf("Error reading body: %v", err)
+	for _, alert := range alerts {
+		log.Infof("Checking alert %s", alert.Id)
+		if alert.IsStatus(common.ALERT_NEW) && alert.OlderThan(ALERT_ESCALATION_TTL) {
+			log.Infof("Escalating alert %s", alert.Id)
+			alert.SetMetadata("status", common.ALERT_ESCALATED)
+
+			// TODO: If we retry based off an error here, we could
+			//		   potentially send an escalation email multiple times.
+			err := globalConfig.sesClient.SendEscalationEmail(alert)
+			if err != nil {
+				log.Errorf("Error escalating alert (%s). Err: %s", alert.Id, err)
+				return err
+			}
+			err = DB.SaveAlert(ctx, alert)
+			if err != nil {
+				log.Errorf("Error updating alert as escalated (%s). Err: %s", alert.Id, err)
+				return err
+			}
+		}
 	}
 
-	// And now set a new body, which will simulate the same data we read:
-	r.Body = ioutil.NopCloser(bytes.NewBuffer(body))
+	return nil
+}
 
-	if cmd, err := slack.SlashCommandParse(r); err == nil && cmd.Command != "" {
-		log.Infof("Command: %s", cmd.Command)
-		if cmd.Command == WHITELIST_IP_SLASH_COMMAND || cmd.Command == STAGING_WHITELIST_IP_SLASH_COMMAND {
-			resp, err := handleWhitelistCmd(r.Context(), cmd, DB)
+func SlackbotBackground(ctx context.Context, psmsg pubsub.Message) error {
+	td, err := common.PubSubMessageToTriggerData(psmsg)
+	if err != nil {
+		log.Errorf("Error decoding pubsub message: %s", err)
+		return nil
+	}
+
+	if td.Action == common.SlashCommand {
+		log.Infof("Got slash command: %s", td.SlashCommand.Cmd)
+		if td.SlashCommand.Cmd == WHITELIST_IP_SLASH_COMMAND || td.SlashCommand.Cmd == STAGING_WHITELIST_IP_SLASH_COMMAND {
+			resp, err := handleWhitelistCmd(ctx, td.SlashCommand, DB)
 			if err != nil {
 				log.Errorf("error handling whitelist command: %s", err)
 			}
 			if resp != nil {
-				err = sendSlackCallback(resp, cmd.ResponseURL)
+				err = sendSlackCallback(resp, td.SlashCommand.ResponseURL)
 				if err != nil {
 					log.Errorf("error sending slack callback within slash command: %s", err)
-					return
+					return err
 				}
 			}
 		}
-	} else if callback, err := InteractionCallbackParse(body); err == nil {
-		if isAlertConfirm(callback) {
-			resp, err := handleAlertConfirm(r.Context(), callback, DB)
+	} else if td.Action == common.Interaction {
+		log.Info("Got interaction action trigger")
+		if isAlertConfirm(td.Interaction.CallbackID) {
+			resp, err := handleAlertConfirm(ctx, td.Interaction, DB)
 			if err != nil {
 				log.Errorf("Error handling alert confirmation interaction: %s", err)
 			}
 			if resp != nil {
-				err = sendSlackCallback(resp, callback.ResponseURL)
+				err = sendSlackCallback(resp, td.Interaction.ResponseURL)
 				if err != nil {
 					log.Errorf("error sending slack callback for interaction: %s", err)
-					return
+					return err
 				}
 			}
 		}
+	} else if td.Action == common.ScheduledTask {
+		log.Info("Got scheduled task action trigger")
+		// We don't want to return an error here, as we don't need the
+		// pubsub/cloudfunction retry mechanism to retry these, as they are scheduled tasks.
+		err = alertEscalator(ctx)
+		if err != nil {
+			log.Errorf("Error escalating alerts: %s", err)
+		}
+		err = DB.RemoveExpiredWhitelistedIps(ctx)
+		if err != nil {
+			log.Errorf("Error purging expired whitelisted ips: %s", err)
+		}
 	}
 
-	return
+	return nil
 }
