+++ slackbot-http/.gcloudignore
