+++ slackbot-http/cmd/main.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"net/http"
+
+	"github.com/mozilla-services/foxsec-pipeline-contrib/slackbot-http"
+)
+
+func main() {
+	http.HandleFunc("/", slackbothttp.SlackbotHTTP)
+	http.ListenAndServe(":8888", nil)
+}
